refactor(timeline): split events mutations into create/lifecycle/helpers (588→157)
events.ts was 588 LOC — all timeline event writes in one file. Extracted verbatim
(behavior-preserving):
- events-helpers.ts — resolveActorId, buildSafeTitle, fetchCreatedEvent,
  handlePostCreationHooks (shared internals)
- events-create.ts — createEventWithVisibility, createEvent, createProjectEvent,
  createTransactionEvent, createQuoteReply

events.ts (157 LOC) keeps the lifecycle ops (updateEvent, updateEventVisibility,
deleteEvent, shareEvent) and re-exports the create ops, so timeline/index.ts keeps
importing all 9 mutations from one place — public surface unchanged.

tsc 0 errors.
